add keys to thumbnails in thumbnail carousel, abstract out getStars, calculateTotalReviewNumber, and calculateAverage from StarReviews.jsx into common/helpers.js, so that the other widgets can utilize them
