+++ src/components/productOverview/StarReviews.jsx
@@ -1,47 +1,19 @@
 /* eslint-disable */
 import React, { useState, useEffect } from 'react';
 import 'whatwg-fetch';
+import { getStars, calculateTotalReviewNumber, calculateAverage } from '../common/helpers.js';
 
 const StarReviews = ({ currentItem, onReviewLinkClick }) => {
   const [rating, setRating] = useState(0);
   const [reviewNum, setReviewNum] = useState(0);
 
-  const calculateAverage = (object) => {
-    let sum = 0;
-    let count = 0;
-    for (const prop in object) {
-      sum += parseInt((prop * object[prop]));
-      count += parseInt(object[prop]);
-    }
-    return Math.round((sum / count) * 10) / 10;
-  };
-
-  const calculateTotalReviewNumber = (object) => {
-    let count = 0;
-    for (const prop in object) {
-      count += parseInt(object[prop]);
-    }
-    return count;
-  };
-
-  const getStars = (productId) => {
-    fetch(`${process.env.API_URI}/reviews/meta?product_id=${productId}`, { method: 'GET', headers: { Authorization: process.env.API_KEY } })
-      .then((response) => {
-        response.json().then((result) => {
-          const average = calculateAverage(result.ratings);
-          const reviews = calculateTotalReviewNumber(result.ratings);
-          setRating(average);
-          setReviewNum(reviews);
-        });
-      })
-      .catch((err) => {
-        console.log(`Error found in getStars: ${err}`);
-      });
-  };
-
   useEffect(() => {
     if (Object.keys(currentItem).length) {
-      getStars(currentItem.id);
+      getStars(currentItem.id)
+      .then((result) => {
+        setRating(result[0]);
+        setReviewNum(result[1])
+      })
     }
   }, [currentItem]);
