+++ src/components/productOverview/ThumbnailCarousel.jsx
@@ -38,9 +38,9 @@ const ThumbnailCarousel = ({photosArray, onThumbnailImageClick, currentStylePhot
       <button className='image-up' onClick={(event) => onThumbnailButtonClick(event)}>Up</button>
       {currentThumbnails.map((photo, i) => {
         if (photo.thumbnail_url === photosArray[currentStylePhotoIndex].thumbnail_url) {
-          return <SelectedThumbnail id={i + count} src={photo.thumbnail_url} onClick={(event) => onThumbnailImageClick(event)}></SelectedThumbnail>
+          return <SelectedThumbnail id={i + count} key={`thumb${i}`} src={photo.thumbnail_url} onClick={(event) => onThumbnailImageClick(event)}></SelectedThumbnail>
         }
-        return <Thumbnail id={i + count} src={photo.thumbnail_url} onClick={(event) => onThumbnailImageClick(event)}></Thumbnail>
+        return <Thumbnail id={i + count} src={photo.thumbnail_url} key={`thumb${i}`} onClick={(event) => onThumbnailImageClick(event)}></Thumbnail>
       })}
       <button className='image-down' onClick={(event) => onThumbnailButtonClick(event)}>Down</button>
     </ThumbnailContainer>
